+++ CHANGES.rst
@@ -19,6 +19,12 @@ unreleased
 - Add IDN support to ``colander.url``.
   See https://github.com/Pylons/colander/pull/352
 
+- ``colander.All`` now supports ``colander.Invalid.msg`` being ``None`` or a
+  list, otherwise ``colander.Invalid.asdict`` crashes with
+  ``TypeError: sequence item 1: expected str instance, NoneType found``.
+  See https://github.com/Pylons/colander/pull/333 and
+  https://github.com/Pylons/colander/issues/194
+
 1.8.3 (2020-11-28)
 ==================
 
@@ -51,6 +57,7 @@ unreleased
   replace then with an ``Invalid`` exception instead.
   See https://github.com/Pylons/colander/pull/338
 
+>>>>>>> master
 1.7.0 (2019-02-01)
 ==================